+++ backend/console/tests/__init__.py
@@ -1,1 +1,2 @@
 from .users import *
+from .workspaces import *
